universal7870: fix audio (overwrite libaudioroute)
we need to use /system/lib instead of /vendor/lib as
libaudioroute is defined by system/media/audio_route
this causes issues with loading audio primary module
and so result in no sound at all.
the REAL fix would be to ensure that AOSP does not
build libaudioroute at all as we use the blob anyways.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>